+++ lnd_methods/signer/index.js
@@ -1,11 +1,17 @@
+const beginGroupSigningSession = require('./begin_group_signing_session');
 const diffieHellmanComputeSecret = require('./diffie_hellman_compute_secret');
+const endGroupSigningSession = require('./end_group_signing_session');
 const signBytes = require('./sign_bytes');
 const signTransaction = require('./sign_transaction');
+const updateGroupSigningSession = require('./update_group_signing_session');
 const verifyBytesSignature = require('./verify_bytes_signature');
 
 module.exports = {
+  beginGroupSigningSession,
   diffieHellmanComputeSecret,
+  endGroupSigningSession,
   signBytes,
   signTransaction,
+  updateGroupSigningSession,
   verifyBytesSignature,
 };